Move HTML util to `utils/`
